Clarified testNoNotesInfoBroadcasting and moved NotebookSocketFake to its own fake class
